Restructure project files and install webpack
