+++ package.json
@@ -2,9 +2,13 @@
   "name": "linear-array",
   "version": "1.0.1",
   "description": "Returns an array filed with linearly increasing numbers, starting from 0 up to the given value - 1 (without offset), or from 1 to the value itself (with offset).",
-  "main": "index.js",
+  "main": "./dist/index.js",
+  "module": "./dist/index.mjs",
+  "files": [
+    "dist"
+  ],
   "scripts": {
-    "test": "mocha"
+    "test": "mocha test/"
   },
   "repository": {
     "type": "git",
@@ -29,10 +33,9 @@
     "url": "https://github.com/ispoljari/linear-array/issues"
   },
   "homepage": "https://github.com/ispoljari/linear-array#readme",
-  "dependencies": {
-    "is-number": "^7.0.0"
-  },
+  "dependencies": {},
   "devDependencies": {
-    "mocha": "^8.0.1"
+    "mocha": "^8.0.1",
+    "webpack": "^4.44.0"
   }
 }
